Struct zeroize::Zeroizing

pub struct Zeroizing<Z: Zeroize>(_);

Zeroizing is a a wrapper for any Z: Zeroize type which implements a Drop handler which zeroizes dropped values.

Implementations

impl<Z> Zeroizing<Z> where
Z: Zeroize,

pub fn new(value: Z) -> Self

Move value inside a Zeroizing wrapper which ensures it will be zeroized when it’s dropped.