Trait zeroize::TryZeroize

pub trait TryZeroize {
    // Required method
    fn try_zeroize(&mut self) -> bool;
}

Fallible trait for representing cases where zeroization may or may not be possible.

This is primarily useful for scenarios like reference counted data, where zeroization is only possible when the last reference is dropped.

Required Methods

fn try_zeroize(&mut self) -> bool

Try to zero out this object from memory using Rust intrinsics which ensure the zeroization operation is not “optimized away” by the compiler.