Trait zeroize::ZeroizeOnDrop

pub trait ZeroizeOnDrop: Zeroize + Drop { }

Marker trait intended for use with zeroize_derive which indicates that a type should have a drop handler which calls Zeroize.

Use #[derive(ZeroizeOnDrop)] to automatically impl this trait and an associated drop handler.