Trait zeroize::Zeroize

pub trait Zeroize {
    fn zeroize(&mut self);
}

Trait for securely erasing types from memory

Required methods

fn zeroize(&mut self)

Zero out this object from memory using Rust intrinsics which ensure the zeroization operation is not "optimized away" by the compiler.

Loading content...